refactor: rename SafePub core modules to private underscore convention
Pure renames plus the matching relative-import line updates, following
the package style of flash/_lattice.py and friends. No logic changes.
